backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt,argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
redis==5.0.4
aioredis==2.0.1
httpx==0.27.0
firebase-admin==6.5.0
apns2==0.7.2
slowapi==0.1.9
cryptography==42.0.7
pyotp==2.9.0
geoip2==4.8.0
python-dotenv==1.0.1
celery==5.4.0
kombu==5.3.4
email-validator==2.1.1
Pillow==10.3.0
qrcode==7.4.2
